refactor(AppShell): uniform vite-plugin naming
