+++ packages/app-shell-vite-plugin/src/tests/vite-plugin-locales.test.ts
