+++ packages/app-shell-vite-plugin/src/vite-plugin-generate-base.ts
